# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickSplit API — bill splitting with StreamPay integration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BillStatus = (typeof BillStatus)[keyof typeof BillStatus];

export const BillStatus = {
  active: "active",
  pending: "pending",
  settled: "settled",
} as const;

export interface Bill {
  id: string;
  payerName: string;
  payerPhone: string;
  description: string;
  totalAmount: string;
  numPeople: number;
  amountPerPerson: string;
  status: BillStatus;
  createdAt: string;
  updatedAt: string;
}

export type SplitItemStatus =
  (typeof SplitItemStatus)[keyof typeof SplitItemStatus];

export const SplitItemStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface SplitItem {
  id: string;
  billId: string;
  amount: string;
  status: SplitItemStatus;
  /** @nullable */
  receiverName?: string | null;
  /** @nullable */
  receiverPhone?: string | null;
  /** @nullable */
  paymentLinkUrl?: string | null;
  /** @nullable */
  streampayConsumerId?: string | null;
  /** @nullable */
  streampayPaymentLinkId?: string | null;
  /** @nullable */
  streampayRef?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type BillWithItemsStatus =
  (typeof BillWithItemsStatus)[keyof typeof BillWithItemsStatus];

export const BillWithItemsStatus = {
  active: "active",
  pending: "pending",
  settled: "settled",
} as const;

export interface BillWithItems {
  id: string;
  payerName: string;
  payerPhone: string;
  description: string;
  totalAmount: string;
  numPeople: number;
  amountPerPerson: string;
  status: BillWithItemsStatus;
  createdAt: string;
  updatedAt: string;
  splitItems: SplitItem[];
}

export interface CreateBillMember {
  name: string;
  phone: string;
}

export interface CreateBillBody {
  payerName: string;
  payerPhone: string;
  description: string;
  totalAmount: number;
  members: CreateBillMember[];
}

export type BillSummaryStatus =
  (typeof BillSummaryStatus)[keyof typeof BillSummaryStatus];

export const BillSummaryStatus = {
  active: "active",
  pending: "pending",
  settled: "settled",
} as const;

export interface BillSummary {
  billId: string;
  description: string;
  totalAmount: string;
  amountPerPerson: string;
  numPeople: number;
  numPaid: number;
  numUnpaid: number;
  amountCollected: string;
  amountRemaining: string;
  status: BillSummaryStatus;
  progressPercent: number;
}

export interface PayerStats {
  totalBills: number;
  activeBills: number;
  pendingBills: number;
  settledBills: number;
  totalCollected: string;
  totalOutstanding: string;
  recentBills: Bill[];
}

export interface JoinBillBody {
  name: string;
  phone: string;
}

export interface JoinBillResponse {
  splitItemId: string;
  billId: string;
  amount: string;
  /** @nullable */
  paymentLinkUrl?: string | null;
  message: string;
}

export type StreamPayWebhookPayloadData = { [key: string]: unknown };

export interface StreamPayWebhookPayload {
  event: string;
  data: StreamPayWebhookPayloadData;
}

export interface WebhookResponse {
  received: boolean;
}

export type ListBillsParams = {
  /**
   * Payer's phone number in E.164 format
   */
  payerPhone: string;
};

export type GetPayerStatsParams = {
  payerPhone: string;
};