# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickSplit API — bill splitting with StreamPay integration
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List bills for a payer
 */
export const ListBillsQueryParams = zod.object({
  payerPhone: zod.coerce
    .string()
    .describe("Payer's phone number in E.164 format"),
});

export const ListBillsResponseItem = zod.object({
  id: zod.string(),
  payerName: zod.string(),
  payerPhone: zod.string(),
  description: zod.string(),
  totalAmount: zod.string(),
  numPeople: zod.number(),
  amountPerPerson: zod.string(),
  status: zod.enum(["active", "pending", "settled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBillsResponse = zod.array(ListBillsResponseItem);

/**
 * @summary Create a new bill with split items
 */
export const CreateBillBody = zod.object({
  payerName: zod.string(),
  payerPhone: zod.string(),
  description: zod.string(),
  totalAmount: zod.number(),
  members: zod.array(
    zod.object({
      name: zod.string(),
      phone: zod.string(),
    }),
  ),
});

/**
 * @summary Get bill details (public, for shared link)
 */
export const GetBillParams = zod.object({
  billId: zod.coerce.string(),
});

export const GetBillResponse = zod.object({
  id: zod.string(),
  payerName: zod.string(),
  payerPhone: zod.string(),
  description: zod.string(),
  totalAmount: zod.string(),
  numPeople: zod.number(),
  amountPerPerson: zod.string(),
  status: zod.enum(["active", "pending", "settled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  splitItems: zod.array(
    zod.object({
      id: zod.string(),
      billId: zod.string(),
      amount: zod.string(),
      status: zod.enum(["unpaid", "paid"]),
      receiverName: zod.string().nullish(),
      receiverPhone: zod.string().nullish(),
      paymentLinkUrl: zod.string().nullish(),
      streampayConsumerId: zod.string().nullish(),
      streampayPaymentLinkId: zod.string().nullish(),
      streampayRef: zod.string().nullish(),
      paidAt: zod.string().nullish(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get payment progress summary for a bill
 */
export const GetBillSummaryParams = zod.object({
  billId: zod.coerce.string(),
});

export const GetBillSummaryResponse = zod.object({
  billId: zod.string(),
  description: zod.string(),
  totalAmount: zod.string(),
  amountPerPerson: zod.string(),
  numPeople: zod.number(),
  numPaid: zod.number(),
  numUnpaid: zod.number(),
  amountCollected: zod.string(),
  amountRemaining: zod.string(),
  status: zod.enum(["active", "pending", "settled"]),
  progressPercent: zod.number(),
});

/**
 * @summary Get aggregated stats for a payer's dashboard
 */
export const GetPayerStatsQueryParams = zod.object({
  payerPhone: zod.coerce.string(),
});

export const GetPayerStatsResponse = zod.object({
  totalBills: zod.number(),
  activeBills: zod.number(),
  pendingBills: zod.number(),
  settledBills: zod.number(),
  totalCollected: zod.string(),
  totalOutstanding: zod.string(),
  recentBills: zod.array(
    zod.object({
      id: zod.string(),
      payerName: zod.string(),
      payerPhone: zod.string(),
      description: zod.string(),
      totalAmount: zod.string(),
      numPeople: zod.number(),
      amountPerPerson: zod.string(),
      status: zod.enum(["active", "pending", "settled"]),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
});

/**
 * @summary Self-register as a member of a bill (via join QR/link)
 */
export const JoinBillParams = zod.object({
  billId: zod.coerce.string(),
});

export const JoinBillBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
});

/**
 * @summary StreamPay payment webhook callback
 */
export const StreampayWebhookBody = zod.object({
  event: zod.string(),
  data: zod.record(zod.string(), zod.unknown()),
});

export const StreampayWebhookResponse = zod.object({
  received: zod.boolean(),
});