# lib/api-zod/src/generated/types/joinBillBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickSplit API — bill splitting with StreamPay integration
 * OpenAPI spec version: 0.1.0
 */

export interface JoinBillBody {
  name: string;
  phone: string;
}