# lib/api-zod/src/generated/types/joinBillResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickSplit API — bill splitting with StreamPay integration
 * OpenAPI spec version: 0.1.0
 */

export interface JoinBillResponse {
  splitItemId: string;
  billId: string;
  amount: string;
  /** @nullable */
  paymentLinkUrl?: string | null;
  message: string;
}